+++ premier-design/features/examples/ui/PhotoViewer/PhotoViewer.tsx
@@ -1,75 +1,20 @@
 'use client';
-import {FC, KeyboardEvent, MouseEvent, useCallback, useEffect, useState} from 'react';
+import {FC} from 'react';
 import styles from './PhotoViewer.module.css';
 import NextImage from 'next/image';
 import {PhotoViewerProps} from "@features/examples/interface/PhotoViewer.props";
+import {usePhotoViewer} from '@features/examples/hooks/usePhotoViewer';
 
 const PhotoViewer: FC<PhotoViewerProps> = ({
                                                images,
                                                currentImage,
                                                onClose
                                            }) => {
-    const [currentIndex, setCurrentIndex] = useState<number>(
-        images.indexOf(currentImage)
-    );
-
-    useEffect(() => {
-        setCurrentIndex(images.indexOf(currentImage));
-    }, [images, currentImage]);
-
-    const handleNext = useCallback(() => {
-        if (images.length > 1) {
-            setCurrentIndex((prevIndex) => (prevIndex + 1) % images.length);
-        }
-    }, [images]);
-
-    const handlePrev = useCallback(() => {
-        if (images.length > 1) {
-            setCurrentIndex((prevIndex) => (prevIndex - 1 + images.length) % images.length);
-        }
-    }, [images]);
-
-    const handleClose = useCallback(() => {
-        onClose();
-    }, [onClose]);
-
-    const handleKeyDown = useCallback(
-        (event: KeyboardEvent<HTMLDialogElement>) => {
-            switch (event.key) {
-                case 'ArrowLeft':
-                    handlePrev();
-                    break;
-                case 'ArrowRight':
-                    handleNext();
-                    break;
-                case 'Escape':
-                    handleClose();
-                    break;
-                case ' ':
-                    if (!event.shiftKey) {
-                        event.preventDefault();
-                        handleNext();
-                    }
-                    break;
-                default:
-                    break;
-            }
-        },
-        [handlePrev, handleNext, handleClose]
-    );
-
-    const handleImageClick = useCallback(
-        (event: MouseEvent<HTMLDivElement>) => {
-            const bounds = event.currentTarget.getBoundingClientRect();
-            const mouseX = event.clientX - bounds.left;
-            if (mouseX < bounds.width / 2) {
-                handlePrev();
-            } else {
-                handleNext();
-            }
-        },
-        [handlePrev, handleNext]
-    );
+    const {currentIndex, handleClose, handleImageClick, handleKeyDown, handleNext, handlePrev} = usePhotoViewer({
+        images,
+        currentImage,
+        onClose,
+    });
 
     return (
         <dialog