+++ premier-design/features/project-offer/hooks/useProjectOffer.ts
@@ -1,0 +1,9 @@
+import {useMemo} from 'react';
+
+type ProjectOfferItem = {
+	id: number | string;
+};
+
+export const useProjectOffer = <T extends ProjectOfferItem>(data: T[]) => {
+	return useMemo(() => data || [], [data]);
+};